infra/requirements.txt
aws-cdk-lib>=2.215.0
constructs>=10.4.2,<11
